api_yamdb/requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==2.2.16
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Faker==8.14.0
flake8==3.9.2
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
idna==2.8
iniconfig==1.1.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.19.1
python3-openid==3.2.0
pytils==0.3
pytz==2019.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.6.3
sqlparse==0.4.2
testfixtures==6.18.0
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==4.1.1
urllib3==1.25.11
wcwidth==0.2.5
zipp==3.5.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
